# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NadlanConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Get the current user's profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/me/role:
    patch:
      tags: [Users]
      operationId: setMyRole
      summary: Set the current user's role (one-time selection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleSelection'
      responses:
        '200':
          description: Role set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
  /users/{userId}:
    patch:
      tags: [Users]
      operationId: adminUpdateUser
      summary: Admin update a user (role, status)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  # ── Listings ──────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, fullName, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        fullName:
          type: [string, 'null']
        role:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        company:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        company:
          type: string
        avatarUrl:
          type: string
    RoleSelection:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [buyer, agent, developer, admin]
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, agent, developer, admin]

    # ── Listings ──
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string