# lib/api-zod/src/generated/types/adminUserUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserUpdateRole } from './adminUserUpdateRole';

export interface AdminUserUpdate {
  role?: AdminUserUpdateRole;
}