# lib/api-zod/src/generated/types/roleSelection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { RoleSelectionRole } from './roleSelectionRole';

export interface RoleSelection {
  role: RoleSelectionRole;
}